fix(arch): merge evaluate_skip_when into fleet._api to stay within file limit
- Inline evaluate_skip_when (and helpers _strip_quotes, _INPUTS_REF_RE)
  into _api.py, reusing the existing _CAMPAIGN_REF_RE defined there;
  eliminates the duplicate-regex cohesion concern as a side effect
- Remove _skip_when.py: fleet/ was at the 15-file sub-package ceiling;
  adding a new file would have failed test_no_subpackage_exceeds_10_files
- Also fixes test_src_uses_regex_not_bare_re: _api.py uses
  'import regex as re', satisfying the arch rule for all src/ modules
- Update fleet/__init__.py to re-export evaluate_skip_when from _api
